# api-gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Examify Swagger",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "paths": {
...
        "/api/ielts/exam/attempt/create/inline": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Creates a new inline attempt for IELTS",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attempts"
                ],
                "summary": "USER",
                "parameters": [
                    {
                        "description": "Create inline attempt request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.CreateInlineAttemptRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    }
                }
            }
        },
        "/api/ielts/exam/attempt/create/outline-speaking/{examId}": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Creates a new inline attempt for IELTS",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attempts"
                ],
                "summary": "USER",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Exam ID",
                        "name": "examId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Question",
                        "name": "question",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Part Number",
                        "name": "partNumber",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Voice Answer (only wav)",
                        "name": "voiceAnswer",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    }
                }
            }
        },
        "/api/ielts/exam/attempt/create/outline-writing": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Creates a new inline attempt for IELTS",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "attempts"
                ],
                "summary": "USER",
                "parameters": [
                    {
                        "description": "Create outline attempt request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pb.CreateOutlineAttemptRequestWriting"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    }
                }
            }
        },
        "/api/ielts/exam/create": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "This endpoint creates a new exam for the specified user and book.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ielts-exam"
                ],
                "summary": "USER",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Book ID",
                        "name": "bookId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Exam created successfully, returning the exam ID",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input parameters",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    }
                }
            }
        },
        "/api/ielts/exam/result/get-results-inline/{sectionType}/{examId}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieves the inline results of a specific section in an IELTS exam",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "results"
                ],
                "summary": "Get results by section",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Section Type",
                        "name": "sectionType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Exam ID",
                        "name": "examId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.GetResultResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    }
                }
            }
        },
        "/api/ielts/exam/result/get-results-outline-speaking/{examId}/{partNumber}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieves the outline results of the Speaking section in an IELTS exam",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "results"
                ],
                "summary": "Get outline of Speaking results",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Exam ID",
                        "name": "examId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Part Number",
                        "name": "partNumber",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.GetResultOutlineSpeakingResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    }
                }
            }
        },
        "/api/ielts/exam/result/get-results-outline-writing/{examId}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieves the outline results of the Writing section in an IELTS exam",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "results"
                ],
                "summary": "Get outline of Writing results",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Exam ID",
                        "name": "examId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/pb.GetResultOutlineWritingResponse"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    }
                }
            }
        },
        "/api/ielts/exam/result/get-results-speaking-voice/{voiceName}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieves the voice record for a given name from the IELTS speaking test",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "audio/wav"
                ],
                "tags": [
                    "results"
                ],
                "summary": "Get a voice record",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Voice Record Name",
                        "name": "voiceName",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns the requested voice file in audio/wav format",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    }
                }
            }
        },
        "/api/ielts/exam/result/top-exam-result/{dataframe}": {
            "get": {
                "description": "Retrieve top exam results based on the specified dataframe (MONTHLY, DAILY, or WEEKLY)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ielts-exam"
                ],
                "summary": "ALL",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The timeframe for which to retrieve top exam results (MONTHLY, DAILY, WEEKLY)",
                        "name": "dataframe",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "The page number for pagination",
                        "name": "page",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "The number of results per page",
                        "name": "size",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with exam results",
                        "schema": {
                            "$ref": "#/definitions/pb.GetTopExamResult"
                        }
                    },
                    "400": {
                        "description": "Bad request with error message",
                        "schema": {
                            "$ref": "#/definitions/utils.AbsResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pb.CreateInlineAttemptRequest": {
            "type": "object",
            "properties": {
                "examId": {
                    "type": "string"
                },
                "sectionType": {
                    "type": "string"
                },
                "userAnswer": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "pb.CreateOutlineAttemptRequestWriting": {
            "type": "object",
            "properties": {
                "Qua": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.QuestionUserAnswerWriting"
                    }
                },
                "examId": {
                    "type": "string"
                }
            }
        },
...
        "pb.GetResultOutlineSpeakingResponse": {
            "type": "object",
            "properties": {
                "coherence_score": {
                    "type": "number"
                },
                "fluency_score": {
                    "type": "number"
                },
                "grammar_score": {
                    "type": "number"
                },
                "part_band_score": {
                    "type": "number"
                },
                "part_number": {
                    "type": "integer"
                },
                "relevance_score": {
                    "type": "number"
                },
                "topic_dev_score": {
                    "type": "number"
                },
                "transcription": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.Transcription"
                    }
                },
                "vocabulary_score": {
                    "type": "number"
                },
                "word_count": {
                    "type": "integer"
                }
            }
        },
        "pb.GetResultOutlineWritingResponse": {
            "type": "object",
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.OutlineWritingResponseAbs"
                    }
                },
                "sectionBandScore": {
                    "type": "integer"
                }
            }
        },
        "pb.GetResultResponse": {
            "type": "object",
            "properties": {
                "answers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.UserAnswer"
                    }
                },
                "bandScore": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                }
            }
        },
        "pb.GetTopExamResult": {
            "type": "object",
            "properties": {
                "results": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/pb.Result"
                    }
                },
                "totalPageCount": {
                    "type": "integer"
                }
            }
        },
        "pb.OutlineWritingResponseAbs": {
            "type": "object",
            "properties": {
                "coherenceScore": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "feedback": {
                    "type": "string"
                },
                "grammarScore": {
                    "type": "integer"
                },
                "lexicalResourceScore": {
                    "type": "integer"
                },
                "question": {
                    "type": "string"
                },
                "taskAchievementScore": {
                    "type": "integer"
                },
                "taskBandScore": {
                    "type": "integer"
                },
                "taskNumber": {
                    "type": "integer"
                },
                "userAnswer": {
                    "type": "string"
                }
            }
        },
        "pb.QuestionUserAnswerWriting": {
            "type": "object",
            "properties": {
                "question": {
                    "type": "string"
                },
                "userAnswer": {
                    "type": "string"
                }
            }
        },
        "pb.Result": {
            "type": "object",
            "properties": {
                "bookName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "examId": {
                    "type": "string"
                },
                "listening": {
                    "type": "string"
                },
                "overall": {
                    "type": "string"
                },
                "reading": {
                    "type": "string"
                },
                "speaking": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/pb.User"
                },
                "writing": {
                    "type": "string"
                }
            }
        },
        "pb.Transcription": {
            "type": "object",
            "properties": {
                "feedback": {
                    "type": "string"
                },
                "question": {
                    "type": "string"
                },
                "transcription": {
                    "type": "string"
                },
                "voiceUrl": {
                    "type": "string"
                }
            }
        },
...
        "pb.User": {
            "type": "object",
            "properties": {
                "chatId": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                },
                "todayExamCount": {
                    "type": "integer"
                }
            }
        },
        "pb.UserAnswer": {
            "type": "object",
            "properties": {
                "isTrue": {
                    "type": "boolean"
                },
                "trueAnswer": {
                    "type": "string"
                },
                "userAnswer": {
                    "type": "string"
                }
            }
        },
        "utils.AbsResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}